Update canMemberAuth to catch auth reason
hasUserAuthed() needed to be refactored to support returning a reason why a user can’t auth.

It’s unorthodox how we check for a reason here instead of, say, “NOT hasAuthed”, but there was no other (easy) way to enable exiting from here when, for example, the user HAS authed but we don’t want to allow them to proceed, or if there’s an error when retrieving the user’s profile.